Restructured folders (links, moved navigation to NavigationController), added search bar
