+++ app/Http/Controllers/UserController.php
@@ -8,24 +8,4 @@ use Illuminate\Http\Request;
 class UserController extends Controller
 {
     //
-
-    public function getIndexPage(){
-        
-    }
-
-    public function getRegisterPage(){
-        $categories = Category::all();
-
-        return view('register', [
-            "categories" => $categories
-        ]);
-    }
-
-    public function getLoginPage(){
-        $categories = Category::all();
-
-        return view('login', [
-            "categories" => $categories
-        ]);
-    }
 }
