+++ public/css/master.css
@@ -1,4 +1,5 @@
 main {
-    background-color: rgb(237, 237, 237);
+    background-color: rgb(218, 218, 218);
     min-height: 92.5vh;
+    padding: 3rem 9rem;
 }
